Refactor 'api' to 'carbon' module
